Added .py on blockchain viz tool
The tool visualizes parsed, queried data from the blockchain.info API to create tree graphs with
simple presentation of addresses and relationships as vertices and edges that helps to reduce the
literacy requirements in understanding the block chain that may have been proven to be a major
challenge to grasp. Supplied parameters act as variants to produce graphs that can cater to specific
flow of funds for a simpler analysis of transactions. Blockchain.info is a fitting reference example
that the general community may use to inspect transactions that involve endless clicks and a manual
paper-and- pen tracing to discover anything meaningful. Such method is unquestionably inefficient
and prone to errors. The tool automates similar exercise by chaining a serious of queries to
immediately produce a graph of relationships with necessary flagging of interested information.